backend/requirements.txt
fastapi==0.105.0
uvicorn==0.25.0
qdrant-client==1.8.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
openai==1.10.0
numpy==1.26.0
pyyaml==6.0.1
python-multipart==0.0.6